Fix re-entrant removal of result listeners in STARTED+
When a fragment result listener is added when the Lifecyle
is already at least STARTED, it is immediately sent the
result when the LifecycleObserver is added. This was
previously done before actually adding the new listener
to the map of result listeners, which meant that
re-entrant clearFragmentResultListener calls wouldn't
find the new listener to remove.

By moving the addObserver after actually adding the
new listener to the map, we ensure that this case works
as expected.

Relnote: "Fixed an issue where calling
`clearFragmentResultListener` inside a
`setFragmentResultListener` wouldn't work if the
`Lifecycle` was already `STARTED` and a result was
already available."
Test: new FragmentResultTest
BUG: 273454803

Change-Id: If7458be3c43ece5a7ee3b1fe5b89bef2cf8b6131