Adjust the timing of OnBackStackChangeCommitted
As currently implemented, onBackStackChangeCommit passes detached
fragments into the callback. Having a fragment with a cleared state is
not very useful.

The more correct timing for this callback would be after the transaction
has been committed but prior to the operations being executed.

RelNote: "The timing of the
`OnBackStackChangedListener.onBackStackChangeCommited` callback has been
adjusted to execute before fragment operations are executed. This
ensures that the callback will never be passed a fragment that is
detached."
Test: Adjusted OnBackStackChagnedListenerTest
Bug: 273568280

Change-Id: I66a76810202a9b44b9d4bdffa03be5ed2b28a1e3