Add ItemSize property to GridTemplate.
The ItemSize refers to a relative bucket size of Small, Medium, or
Large. The size is applied to all items in the grid. The default size is small. ItemSize will impact the grid item
image size and the width of each grid item, as well as the number of
items displayed per row. These values will also depend on the screen
size.

This change is experimental for now, and will be part of
API level 7.

Test: unit tests
Bug: 271862896
Relnote: "Add ItemSize property to GridTemplate, which controls grid
item sizing according to relative small, medium, large buckets."

Change-Id: Icdb3bccfc7e9f72d580a9d3e88ce306ce3edbb08