Address Grid Dsl Api Council comments
This CL is created to address two comments from the API Council
1.Should paddingLeft and paddingRight be paddingStart and paddingEnd?
2.Why is gridSpans and gridSkips a string? Maybe we should do validation close to the top of the call stack

For 1, update paddingLeft and paddingRight to paddingStart and paddingEnd, respectively
For 2, update the format of gridSpans and gridSkip to an array of the Span and Skip objects to make it more structural.

Bug: 273543249
Bug: 273951887
Test: pass GridDslTest.kt and RowColumnDslTest.kt
Relnote: Two changes made for the Grid helper: 1. paddingLeft and paddingRight to paddingStart and paddingEnd, respectively, and 2. update the format of gridSpans and gridSkip tos an array of the Span and Skip objects to make it more structural.
Change-Id: Idd1ebe78c874c6d40e6e39e4d93819061c80fe6b