Added support to test different gradle versions
This PR adds support to the `BaselineProfileProjectSetupRule` to force
a specific agp version when running the plugin. This is done removing
the default plugin class to the gradle runner (that is the classpath
of the test running) and specifying instead as classpath configurations
in the root build.gradle both agp version and plugin.

Test: ./gradlew :benchmark:integration-tests:baselineprofile-library-consumer:testBaselineProfileGeneration :benchmark:integration-tests:baselineprofile-flavor-consumer:testBaselineProfileGeneration :benchmark:integration-tests:baselineprofile-consumer:testBaselineProfileGeneration :benchmark:benchmark-baseline-profile-gradle-plugin:test
Bug: 273826398
Change-Id: If3e38272871794b7a544475cacd8376f71130936