Implement LazyPagingItem itemKey as helper function
Added itemKey API which returns a lambda that can be plugged into `key` params for Lazy scopes.
If no key is provided, defaults to PagingPlaceholderKey(index) instead of compose's DefaultLazyKey

Test: Compose demos
Test: ./gradlew paging:paging-compose:cC
Bug: 259385813
Relnote: "LazyPagingItems has a new itemKey API which returns a key implementation that can be plugged into the `key` params of Lazy scopes. If no key is provided, defaults to using index-based PagingPlaceholderKey."

Change-Id: Ifa13b3d764203422645cac8bfc3d09278d508874