Address feedback on AndroidView API
This CL un-deprecates the original overload of `AndroidView` and adds
a KDoc back to the function. By doing so and keeping both functions
alive, we can make sure that all existing usages of AndroidView
continue to function when compiled against newer versions of the UI
library, since Kotlin will prefer the existing overload with fewer
arguments than the new overloaded with extra callbacks.

Relnote: N/A
Test: N/A
Fixes: b/273287604
Change-Id: I17a9fcc74137f214317de289ee9ff17730b4e19a