Store generated sources from tests in build output.
Currently our golden diff tests store the generated files in a temporary
directory that is soon deleted after the test finishes. This makes
changing multiple goldens at once significantly harder as we need to
copy and paste from the test output.

With this change we are storing the generated test files in the standard
build output. The diff assertion will print out the path for the files,
so we can later copy them if they are what we expect.

As a bonus, I also introduced a common abstract test class for our diff
testing and removed duplicate logic from the API Compiler tests, like we
did with the generator in the past.

Test: ./gradlew :privacysandbox:tools:tools-{apigenerator,apicompiler,core,apipackager}:test --parallel
Change-Id: I3de213daaa3a421bfc79ce31e45a22ac2211bcc2