Revert "Revert "[CameraPipe] Add lock behaviors for all 3A modes for startFocusAndMetering""
This reverts commit 1929742186a62095314ae21441465669c335d7ba.

Reason for revert:
It seems that re-applying the reverted CL on LKGB finds no flakiness (result in b/274588483). So, the bug is probably not related with the reverted CL and was caused/fixed by something else.

Also, the whole AudioSourceTest has been ignored temporarily in aosp/2504915 and the root cause is being investigated in b/274840083

Change-Id: If4747d00aeec6c5b4627337429dea84f3cc5820a