[CredMan] Prefetch jetpack apis matching the new framework changes.
This change contains the APIs matching the new Credential Manager
prefetch apis in the framework (b/273308895).

Note 1: I can't do the actual implementation because the framework
changes have not been propagated to the proper release branch. Will send
a follow-up implementation as soon as unblocked.

Note 2: the prefetch APIs requires U+ for now but is designed so that it
can be exposed to <U in the future if we get such a feature request.

Relnote: "Credential Manager prefetch apis"
Bug: 274976698
Test: unit tested
Change-Id: I1c362262dccddc678c7e20648977aad6f6b03cbe