Add PeriodicWorkRequest.setNextScheduleTimeOverride
This feature allows developers to accurately set the next time a
periodic worker will run.

setNext adds an additional field to WorkSpec which overrides
any other schedule calculation. It will clear itself after the next time
the work runs. Other fields are maintained, so if the override time is
cleared explicitly, scheduling will revert to whatever it would have been before
based on backoff, period, etc.

A critical usecase is using setNext from within a Periodic Worker
to set its own next run time. This is provided for by use of a
generation counter. WorkerWrapper will only clear the next runtime if
the override generation hasn't changed. Otherwise, it will leave the
override for the next run.

The minimum period interval (15 minutes) is respected for non-initial
periods, to prevent infinite immediate loops. For this reason,
ExistingPeriodicWorkPolicy.CANCEL_AND_REENQUEUE is not allowed to be
used, since it would reset the periodCount to 0 and prevent this check
from working.

Test: Added many tests for this feature.
Relnote: "Added setNextScheduleTimeOverride method, which allows accurate
setting of periodic work schedules"

Change-Id: I3b4da91bab9ffc6da3e305ef7ae814b59f7090fb