Fix robolectric test IllegalArgumentException issue in SupportedSurfaceCombination when updating the library artifacts
App might have robolectric tests but does not setup output sizes for all formats that might be used. The robolectric test might only create the CameraX instance but not bind use cases to a lifecycle owner. In that case, output sizesmight be unnecessary to be setup for the test environment. Update the code to query the output sizes when it is really necessary when binding use cases.

Added a test in camera-camera2 to make sure that SupportedSurfaceCombnation instance can be created when no StreamConfigurationMap and output sizes are set.

For camera-camera2-pipe-integration, it will fail to create a CameraX instance when no StreamConfigurationMap is set. Need further clarification but it should not block the issue.

Bug: 275618995
Test: ./gradlew bOS
Change-Id: I4824ba37504cc9aa853b5cab35b7a32899a0457e