Add a quirk that waits for onClosed() during camera close
On legacy camera devices, releasing a Surface while still producing
frames causes an Exception. Here we add a quirk that handles it by
waiting for onClosed() while we're closing the camera.

We also fix a bug where, if multiple CaptureSessionState.disconnect() is
called in parallel, we prematurely and erroneously trigger the path to
finalize the session as if there was no cameras ever opened at all.

Bug: None
Test: CameraPipe unit tests
Change-Id: Ic048107b2d03463aa4ad62fc948294a098461293