Fix potential deadlock in snapshot list and map
Adding content to a SnapshotStateList or SnapshotStateMap can
encounter a deadlock if the modification is concurrent with
a direct write to the state record. This was made signficantly
more likely to be encountered  with the changes introduced by
c2a1fbaeec that uses direct writes to release unused records.

The locks are now ordered in that a snapshot lock is never
attempted to be taken when a map or list lock is held.

Fixes: 269402895
Test: ./gradlew :composer:r:r:tDUT
(cherry picked from https://android-review.googlesource.com/q/commit:aeb8125937cfb1a76cd7be11c136a03fed8aeb54)
Merged-In: Ie8b87caecabc88f6a6dd610805d964aed3489b85
Change-Id: Ie8b87caecabc88f6a6dd610805d964aed3489b85