Ensure postponed with timeout does not cause leak
If you use postponeEnterTransition with a timeout, we add a runnable to
do the handler to ensure we eventually call
startPostponedEnterTransition, but we don't clear that runnable from the
handler when we do startPostponedEnterTransition. This means we continue
to have a reference to the runnable and causes the fragment to leak.

We should ensure we remove the runnable from the handler when we
startPostponedEnterTransition.

RelNote: "Using `postponeEnterTransition` with a timeout and then
replace the postponed fragment no longer results in leaking the
postponed fragment."
Test: added postponedTest
Bug: 276375110

Change-Id: I2ec7d50c9a08be4902dcc3cce4615bc1e19ed47d