Add onReset and onRelease to AndroidViewBinding
`AndroidViewBinding` is a Composable that builds on top of
`AndroidView` for users wanting compatibility with existing layouts
that use ViewBinding. We also have advice in places to use
AndroidViewBinding to host Fragments.

This CL updates AndroidViewBinding to include the `onReset` and
`onRelease` parameters, mirroring `AndroidView`. This allows users to
opt-in to View and binding reuse for both variations of AndroidView
by providing a non-null `onReset` parameter, and to observe View
disposal with the `onRelease` parameter.

Relnote: """Added `onReset` and `onRelease` parameters to the
    `AndroidViewBinding` composable, mirroring the `AndroidView`
    composable and enabling support for View reuse with ViewBinding.
    """
Test: Updated samples and tests in ui-viewbinding module
Bug: b/276802519
Change-Id: I00b1e2a22dde29b6e011d3d8bb1ea51391231856