Configuration Change Consistency in Play Auth Flow
This change is a follow up to a prior change that handles more general
configuration changes - or really anything that restarts the app
(assuming it is run in the proper scope, namely a non cancellation
inducing scope in the programming). This ensures that the behaviour of the underlying
module is consistent and returns credentials as expected.

Bug: 276316128
Bug: 276939495
Test: E2E Test and Build
Change-Id: Iae8ce99fa56d237d874c5af2c61a914aaaeb85e4