Add contentKey support to AnimatedContent
This CL adds the contentKey mapping to AnimatedContent.
Note, this lambda has been in `Transition.AnimatedContent`
for a while because we were holding off adding it to the
higher traffic API due to the concern that it might cause
confusion. From recent discussions, devs expressed confusion
around why contentKey is not present in all AnimatedContent APIs.
It also seems that the use case that contentKey is serving
is more common than we expected, given the times the discussion
or question around this comes up.

Also updated the default animationSpec for SizeTransform.

Test: New test added.
RelNote: "
ContentKey is now supported in AnimatedContent to allow more
control for when transition should happen.
"

Change-Id: Ic069b01c0d2b302c7527ed850c32ba68bfc0101f