Add LinkMovementMethodCompat that fixes link touch handling.
LinkMovementMethod incorrectly reacted to touches outside of line
bounds, which is considered an accessibility bug and fixed in Android V
in ag/22301465. It would be helpful for apps to have a
backwards-compatible version of the fix inside AndroidX.

Note that we can't simply copy over the entire LinkMovementMethod
class, because it contains references to hidden APIs. Instead of that,
we can simply override onTouchEvent() and handle outside touches
differently but leave the rest to the original implementation.

Bug: 271473145
Test: LinkMovementMethodCompatTest
Relnote: "Added `LinkMovementMethodCompat` that fixes link touch
    handling."
Change-Id: Ia632c50bc335cf42ab903a2521de8afd97737f83