Keep the sizes for TargetAspectRatio workaround in the end of the list in SupportedSurfaceCombination
It is not guaranteed that at least one output size can be kept in VGA, PREVIEW, RECORD, MAXIMUM size levels if we directly filtering output sizes according to the TargetAspectRatio workaround. Therefore, rollback to keep sizes (as what it is before aosp/2442944 was merged) which do not meet the TargetAspectRatio requirements to be in the end of the sizes list in SupportedSurfaceCombination. Those sizes can still be selected in lower priority in case the devices might not have available size of the specific aspect ratio in a small size level for a UseCase.

Bug: 276483197
Test: ./gradlew bOS

Change-Id: I3a172139d2553f33817c61d4fcd12a3220d9713e