Sample and RevealScope for SwipeToReveal
Add `RevealScope` for actions in SwipeToReveal. It will expose
the common properties like offset values for the actions
composable to use.

Added some samples for the SwipeToReveal

Bug: 277871563
Bug: 277870075
Bug: 278717383
Bug: 278717643
Bug: 278041206
Bug: 278013751
Bug: 278255008
Test: SwipeToRevealTest
Relnote: """
RevealScope for action composables in SwipeToReveal. The scope
object will allow action composables to customise the content
based on different anchors and total width.
"""

Change-Id: I3fd56eb7f8f5f9ec467790988afb9f942766098d