Add IconButton to Material3
IconButton is a circular button with a single slot for icon/image. There are four variations: IconButton, FilledIconButton, FilledTonalIconButton and OutlinedIconButton. For different sizes of IconButton, we recommend using Modifier.touchTargetAwareSize and ExtraSmallButtonSize, SmallButtonSize, DefaultButtonSize and LargeButtonSizeIcon provided in IconButtonDefaults. We also provide IconButtonDefaults.iconSizeFor to determine the recommended icon size for a given button size.

Test: N/A
Bug: 278713522
Relnote: "We have added IconButton to Material3, a circular button with a single slot for icon/image. There are four variations: IconButton, FilledIconButton, FilledTonalIconButton and OutlinedIconButton. For different sizes of IconButton, we recommend using Modifier.touchTargetAwareSize and ExtraSmallButtonSize, SmallButtonSize, DefaultButtonSize and LargeButtonSizeIcon provided in IconButtonDefaults. We also provide IconButtonDefaults.iconSizeFor to determine the recommended icon size for a given button size."
Change-Id: I721d46af2b1d2d17185cbcf530e831f03a53e47e