Fix some usages of System.currentTimeMillis() I missed.
There remains a couple usages still (eg. in WorkRequestBuilder) which I
couldn't easily pipe the clock to. It would possibly result in a
requirement for WM to be initialized sooner than it currently is, which
I don't want to do unless I have to.

For now substituting the clock will be a test-only ability to avoid the
remaining implicit System clock from conflicting with a user-configured
clock in prod.

Test: Existing tests pass.
Change-Id: Ic7a0fd946eb32963538c01b28db9a9254c8c42af