fix core-telecom test failures by adding SdkSuppress annotation
Test breakages came in for all the androidx.core.telecom tests. Upon
investigation these tests were failing on a device with an sdk=17. The
entire test unit module should only be run against sdks=26 and above.
The SdkSuppress annotation is able to guarantee this.

Fixes: 279741430
RelNote: add the SdkSuppress annotation to core-telecom tests
Test: existing unit tests in androidx.core.telecom
Change-Id: Ie65b6698bcd9ae7d01b04c01a54c7718241a1acf