Poll last ActivityResult request
During registry of the ActivityResultRegistry in
FragmentManager, we always polled the first
startActivityForResult request. However, if users
happen to have launched more than one request
consecutively, this will cause us to grab the
wrong request information for the ActivityResult
we are receiving, since results would dispatch
to onActivityResult in reverse order of when the
request was made.

Note: The original bug results from using the
**deprecated** APIs. Please be sure to use
the Activity Result APIs instead.

RelNote: "Bug fixed causing `ActivityResult`s
to be sent with the incorrect request code when
multiple `startActivityForResult` requests have
been made consecutively."
Fixes: 249519359
Test: launchMultipleActivitiesFromFragment

Change-Id: If0b9da0f226144df6c9ce0ef76642cf47184e63c