Followup: fix preview rotation in core test app
This CL first reverts the vertex coordinates change made in aosp/2549510. Then I observed the fact that in #getViewPortRotation() we were adding surface rotation with texture rotation which is incorrect. Surface rotation and texture rotation should offset each other, instead of adding on each other.

So I made the fix by replacing mTextureRotationDegrees with (180 - mTextureRotationDegrees). My assuption is that the 180 compensates for the shader canvas rotation.

Verified when front/back camera with video on/off, orientation lock on/off.

Bug: 269613528
Test: manual test and ./gradlew bOS
Change-Id: I599ee72308ba8770ad10fa2e2067ea2ebe99b49b