Fix native crash due to incorrect API level check
When the release name changes for a platform build, the
isAtLeastXXX() code that previously worked make not work
correctly. This means that code using that check will
end up in a code path intended for a different platform level.

In the case of this library, that meant ending up in native code
which made an assumption about the data structures in platform
code that are no longer correct in API 34, so it crashed.

The fix was to use the ToT isAtLeastU() version (instead of whatever
is in prebuilts), since the ToT version is updated quickly after the
platform change happens.

Bug: 280639559
Test: PathIteratorTests pass
Change-Id: I2ca975c82106b64c15066d1992b33af0461edad4